requirements.txt
appdirs==1.4.3
Django==2.0.0
django-appconf==1.0.2
django-bower==5.2.0
django-filter==1.0.4
django-imagekit==4.0.1
django-material==1.3.0
djangorestframework==3.8.0
Faker==0.7.12
ipaddress==1.0.18
Markdown==2.6.8
numpy==1.12.1
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
